date localization tests, partly done
